rename bapsflib.utils.errors -> bapsflib.utils.excpetions
