+++ bapsflib/_hdf/maps/controls/nixz.py
@@ -17,7 +17,7 @@ __all__ = ["HDFMapControlNIXZ"]
 import h5py
 import numpy as np
 
-from bapsflib.utils.errors import HDFMappingError
+from bapsflib.utils.exceptions import HDFMappingError
 from warnings import warn
 
 from .types import ConType
