+++ bapsflib/_hdf/maps/controls/tests/test_nixz.py
@@ -14,7 +14,7 @@
 import h5py
 import unittest as ut
 
-from bapsflib.utils.errors import HDFMappingError
+from bapsflib.utils.exceptions import HDFMappingError
 
 from .common import ControlTestCase
 from .. import ConType
