+++ bapsflib/_hdf/maps/controls/tests/test_sixk.py
@@ -15,7 +15,7 @@ import numpy as np
 import os
 import unittest as ut
 
-from bapsflib.utils.errors import HDFMappingError
+from bapsflib.utils.exceptions import HDFMappingError
 from unittest import mock
 
 from .common import ControlTestCase
