+++ bapsflib/_hdf/maps/digitizers/map_digis.py
@@ -13,7 +13,7 @@ __all__ = ["HDFMapDigitizers"]
 
 import h5py
 
-from bapsflib.utils.errors import HDFMappingError
+from bapsflib.utils.exceptions import HDFMappingError
 from typing import (Dict, Tuple)
 
 from .sis3301 import HDFMapDigiSIS3301
