+++ bapsflib/_hdf/maps/digitizers/sis3301.py
@@ -20,7 +20,7 @@ import numpy as np
 import os
 import re
 
-from bapsflib.utils.errors import HDFMappingError
+from bapsflib.utils.exceptions import HDFMappingError
 from typing import (Any, Dict, Tuple, Union)
 from warnings import warn
