+++ bapsflib/_hdf/maps/msi/tests/test_discharge.py
@@ -14,7 +14,7 @@
 import numpy as np
 import unittest as ut
 
-from bapsflib.utils.errors import HDFMappingError
+from bapsflib.utils.exceptions import HDFMappingError
 
 from ..discharge import HDFMapMSIDischarge
 from .common import MSIDiagnosticTestCase
