+++ bapsflib/_hdf/maps/msi/tests/test_interferometerarray.py
@@ -14,7 +14,7 @@
 import numpy as np
 import unittest as ut
 
-from bapsflib.utils.errors import HDFMappingError
+from bapsflib.utils.exceptions import HDFMappingError
 
 from ..interferometerarray import HDFMapMSIInterferometerArray
 from .common import MSIDiagnosticTestCase
