+++ bapsflib/utils/exceptions.py
